Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit a3cb76973c58496424e893837152b192e77bb025, reversing
changes made to 1d9b497fc8437aafc674c12cfe12d1556af28069.

+++ code/modules/projectiles/projectile/special/wormhole.dm
@@ -2,15 +2,9 @@
 	name = "bluespace beam"
 	icon_state = "spark"
 	hitsound = "sparks"
-	damage = 0
-	nodamage = TRUE
-	pass_flags = PASSGLASS | PASSTABLE | PASSGRILLE | PASSMOB
+	damage = 3
 	var/obj/item/gun/energy/wormhole_projector/gun
 	color = "#33CCFF"
-	tracer_type = /obj/effect/projectile/tracer/wormhole
-	impact_type = /obj/effect/projectile/impact/wormhole
-	muzzle_type = /obj/effect/projectile/muzzle/wormhole
-	hitscan = TRUE
 
 /obj/item/projectile/beam/wormhole/orange
 	name = "orange bluespace beam"
@@ -21,9 +15,11 @@
 	if(casing)
 		gun = casing.gun
 
-
 /obj/item/projectile/beam/wormhole/on_hit(atom/target)
+	if(ismob(target))
+		var/turf/portal_destination = pick(orange(6, src))
+		do_teleport(target, portal_destination)
+		return ..()
 	if(!gun)
 		qdel(src)
-		return 
 	gun.create_portal(src, get_turf(src))
